requirements.txt
blinker==1.4
Flask==0.10.1
Flask-Mail==0.9.1
Flask-MySQL==1.3
Flask-MySQLdb==0.2.0
Flask-Sendmail==0.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
mysqlclient==1.3.6
pycrypto==2.6.1
simple-crypt==4.1.7
Werkzeug==0.10.4
wheel==0.24.0
